# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/permissions": {
      "get": {
        "operationId": "RolesController_permissions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Grantable permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionOptionsEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Permissions the caller may put into a role (the catalog it already holds)",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles": {
      "get": {
        "operationId": "RolesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A page of roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List roles: the shared system roles and the tenant's own",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "RolesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a custom role in a tenant",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "operationId": "RolesController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one role with its permissions",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_rename_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Rename a custom role (system roles are immutable)",
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove_v1",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a custom role that is not assigned to anyone",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}/permissions": {
      "put": {
        "operationId": "RolesController_setPermissions_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The role with its new permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace the permissions of a custom role (never more than you hold)",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/tenants": {
      "get": {
        "operationId": "TenantsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A page of tenants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List tenants visible to the caller (own tenant; all for platform admin)",
        "tags": [
          "tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a tenant (platform admin only)",
        "tags": [
          "tenants"
        ]
      }
    },
    "/api/v1/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one tenant",
        "tags": [
          "tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a tenant (platform admin may also activate or deactivate)",
        "tags": [
          "tenants"
        ]
      }
    },
...
  },
  "info": {
    "title": "Starter App API",
    "description": "Success responses are wrapped in `{ data, meta? }`; errors follow RFC 9457 Problem Details with `application/problem+json`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token only. Swagger adds the Bearer prefix."
      }
    },
    "schemas": {
...
      "PermissionOptionsEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "presets": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "any",
                      "own_tenant",
                      "own_record"
                    ]
                  }
                }
              },
              "required": [
                "action",
                "subject",
                "description",
                "presets"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "RoleListEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "key": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "scope": {
                  "type": "string",
                  "enum": [
                    "platform",
                    "tenant"
                  ]
                },
                "isSystem": {
                  "type": "boolean"
                },
                "tenantId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "permissions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "enum": [
                          "manage",
                          "create",
                          "read",
                          "update",
                          "delete"
                        ]
                      },
                      "subject": {
                        "type": "string"
                      },
                      "preset": {
                        "type": "string",
                        "enum": [
                          "any",
                          "own_tenant",
                          "own_record"
                        ]
                      }
                    },
                    "required": [
                      "action",
                      "subject",
                      "preset"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "key",
                "name",
                "scope",
                "isSystem",
                "tenantId",
                "permissions"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "permissions": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "enum": [
                    "create",
                    "read",
                    "update",
                    "delete"
                  ]
                },
                "subject": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 60
                },
                "preset": {
                  "type": "string",
                  "enum": [
                    "any",
                    "own_tenant",
                    "own_record"
                  ]
                }
              },
              "required": [
                "action",
                "subject",
                "preset"
              ]
            }
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "RoleEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "key": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "scope": {
                "type": "string",
                "enum": [
                  "platform",
                  "tenant"
                ]
              },
              "isSystem": {
                "type": "boolean"
              },
              "tenantId": {
                "type": "string",
                "format": "uuid",
                "nullable": true
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "action": {
                      "type": "string",
                      "enum": [
                        "manage",
                        "create",
                        "read",
                        "update",
                        "delete"
                      ]
                    },
                    "subject": {
                      "type": "string"
                    },
                    "preset": {
                      "type": "string",
                      "enum": [
                        "any",
                        "own_tenant",
                        "own_record"
                      ]
                    }
                  },
                  "required": [
                    "action",
                    "subject",
                    "preset"
                  ]
                }
              }
            },
            "required": [
              "id",
              "key",
              "name",
              "scope",
              "isSystem",
              "tenantId",
              "permissions"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          }
        },
        "required": [
          "name"
        ]
      },
      "SetRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "enum": [
                    "create",
                    "read",
                    "update",
                    "delete"
                  ]
                },
                "subject": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 60
                },
                "preset": {
                  "type": "string",
                  "enum": [
                    "any",
                    "own_tenant",
                    "own_record"
                  ]
                }
              },
              "required": [
                "action",
                "subject",
                "preset"
              ]
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "TenantListEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "isActive": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "name",
                "slug",
                "isActive",
                "createdAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "maxLength": 60
          }
        },
        "required": [
          "name"
        ]
      },
      "TenantEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "isActive": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "name",
              "slug",
              "isActive",
              "createdAt"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}